refactor: extract directory proxy
